refactor: Extract Service logic from App into a hook
App shouldn't be in charge of running the service.
